+++ altair/utils/_transformed_data.py
@@ -25,7 +25,7 @@ from altair import (
     data_transformers,
 )
 from altair.utils._vegafusion_data import get_inline_tables, import_vegafusion
-from altair.utils.core import _DataFrameLike
+from altair.utils.core import DataFrameLike
 from altair.utils.schemapi import Undefined
 
 Scope = Tuple[int, ...]
@@ -56,7 +56,7 @@ def transformed_data(
     chart: Union[Chart, FacetChart],
     row_limit: Optional[int] = None,
     exclude: Optional[Iterable[str]] = None,
-) -> Optional[_DataFrameLike]:
+) -> Optional[DataFrameLike]:
     ...
 
 
@@ -65,7 +65,7 @@ def transformed_data(
     chart: Union[LayerChart, HConcatChart, VConcatChart, ConcatChart],
     row_limit: Optional[int] = None,
     exclude: Optional[Iterable[str]] = None,
-) -> List[_DataFrameLike]:
+) -> List[DataFrameLike]:
     ...