+++ altair/utils/data.py
@@ -10,7 +10,7 @@ from toolz import curried
 from typing import TypeVar
 
 from ._importers import import_pyarrow_interchange
-from .core import sanitize_dataframe, sanitize_arrow_table, _DataFrameLike
+from .core import sanitize_dataframe, sanitize_arrow_table, DataFrameLike
 from .core import sanitize_geo_interface
 from .deprecation import AltairDeprecationWarning
 from .plugin_registry import PluginRegistry
@@ -23,15 +23,15 @@ if TYPE_CHECKING:
     import pyarrow.lib
 
 
-class _SupportsGeoInterface(Protocol):
+class SupportsGeoInterface(Protocol):
     __geo_interface__: MutableMapping
 
 
-_DataType = Union[dict, pd.DataFrame, _SupportsGeoInterface, _DataFrameLike]
-_TDataType = TypeVar("_TDataType", bound=_DataType)
+DataType = Union[dict, pd.DataFrame, SupportsGeoInterface, DataFrameLike]
+TDataType = TypeVar("TDataType", bound=DataType)
 
-_VegaLiteDataDict = Dict[str, Union[str, dict, List[dict]]]
-_ToValuesReturnType = Dict[str, Union[dict, List[dict]]]
+VegaLiteDataDict = Dict[str, Union[str, dict, List[dict]]]
+ToValuesReturnType = Dict[str, Union[dict, List[dict]]]
 
 
 # ==============================================================================
@@ -46,7 +46,7 @@ _ToValuesReturnType = Dict[str, Union[dict, List[dict]]]
 # form.
 # ==============================================================================
 class DataTransformerType(Protocol):
-    def __call__(self, data: _DataType, **kwargs) -> _VegaLiteDataDict:
+    def __call__(self, data: DataType, **kwargs) -> VegaLiteDataDict:
         pass
 
 
@@ -70,7 +70,7 @@ class MaxRowsError(Exception):
 
 
 @curried.curry
-def limit_rows(data: _TDataType, max_rows: Optional[int] = 5000) -> _TDataType:
+def limit_rows(data: TDataType, max_rows: Optional[int] = 5000) -> TDataType:
     """Raise MaxRowsError if the data model has more than max_rows.
 
     If max_rows is None, then do not perform any check.
@@ -122,7 +122,7 @@ def limit_rows(data: _TDataType, max_rows: Optional[int] = 5000) -> _TDataType:
 
 @curried.curry
 def sample(
-    data: _DataType, n: Optional[int] = None, frac: Optional[float] = None
+    data: DataType, n: Optional[int] = None, frac: Optional[float] = None
 ) -> Optional[Union[pd.DataFrame, Dict[str, Sequence], "pyarrow.lib.Table"]]:
     """Reduce the size of the data model by sampling without replacement."""
     check_data_type(data)
@@ -180,7 +180,7 @@ class _ToCsvReturnUrlDict(TypedDict):
 
 @curried.curry
 def to_json(
-    data: _DataType,
+    data: DataType,
     prefix: str = "altair-data",
     extension: str = "json",
     filename: str = "{prefix}-{hash}.{extension}",
@@ -199,7 +199,7 @@ def to_json(
 
 @curried.curry
 def to_csv(
-    data: Union[dict, pd.DataFrame, _DataFrameLike],
+    data: Union[dict, pd.DataFrame, DataFrameLike],
     prefix: str = "altair-data",
     extension: str = "csv",
     filename: str = "{prefix}-{hash}.{extension}",
@@ -215,7 +215,7 @@ def to_csv(
 
 
 @curried.curry
-def to_values(data: _DataType) -> _ToValuesReturnType:
+def to_values(data: DataType) -> ToValuesReturnType:
     """Replace a DataFrame by a data model with values."""
     check_data_type(data)
     if hasattr(data, "__geo_interface__"):
@@ -242,7 +242,7 @@ def to_values(data: _DataType) -> _ToValuesReturnType:
         raise ValueError("Unrecognized data type: {}".format(type(data)))
 
 
-def check_data_type(data: _DataType) -> None:
+def check_data_type(data: DataType) -> None:
     if not isinstance(data, (dict, pd.DataFrame)) and not any(
         hasattr(data, attr) for attr in ["__geo_interface__", "__dataframe__"]
     ):
@@ -260,7 +260,7 @@ def _compute_data_hash(data_str: str) -> str:
     return hashlib.md5(data_str.encode()).hexdigest()
 
 
-def _data_to_json_string(data: _DataType) -> str:
+def _data_to_json_string(data: DataType) -> str:
     """Return a JSON string representation of the input data"""
     check_data_type(data)
     if hasattr(data, "__geo_interface__"):
@@ -288,7 +288,7 @@ def _data_to_json_string(data: _DataType) -> str:
         )
 
 
-def _data_to_csv_string(data: Union[dict, pd.DataFrame, _DataFrameLike]) -> str:
+def _data_to_csv_string(data: Union[dict, pd.DataFrame, DataFrameLike]) -> str:
     """return a CSV string representation of the input data"""
     check_data_type(data)
     if hasattr(data, "__geo_interface__"):